Creates an object ready for use in scX

createSCEobject creates the input object (as a List) for the function launch_scX. This list includes a SingleCellExperiment object with normalized expression, reduced dimensions for visualization, and any additional data provided. Also, it includes gene markers and differential expression analysis for each specified partition (if no partition is selected, a quick clusterization is computed)

Usage

createSCEobject(
  xx,
  assay.name.raw = "counts",
  assay.name.normalization = "logcounts",
  metadata = NULL,
  partitionVars = NULL,
  metadataVars = NULL,
  chosen.hvg = NULL,
  nHVGs = 3000,
  nPCs = 50,
  calcRedDim = TRUE,
  paramFindMarkers = list(test.type = "wilcox", pval.type = "all", direction = "any"),
  calcAllPartitions = FALSE,
  cells2keep = NULL,
  descriptionText = NULL,
  verbose = TRUE
)

Arguments

xx

Either a numeric matrix-like object of the raw count matrix (with genes as rows and cells as columns), a SingleCellExperiment object or Seurat object.

assay.name.raw

Assay name for raw counts matrix if object is a SingleCellExperiment. Defaults to counts.

assay.name.normalization

Assay name for normalized matrix if present in SingleCellExperiment. If not present, it computes logcounts (default).

metadata

Optional DataFrame with cell metadata. Rownames in the metadata must include all cell names of xx.

partitionVars

Optional metadata column names (or colData) to use for gene markers and differential expression analysis. If NULL (default), a quick clusterization using quickCluster from scran package will be computed.

metadataVars

Additional metadata column names (or colData)) to use only for coloring in plots. If NULL (default), only partitionVars columns will be available for coloring plots.

chosen.hvg

Optional list of Highly Variable Genes. NOTE: if chosen.hvg=NULL and xx is a Seurat object with computed VariableFeatures then this parameter will be set to that list of genes.

nHVGs

Number of Highly Variable Genes to use if chosen.hvg=NULL. Defaults to 3000.

nPCs

Number of Principal Components to use in PCA. Defaults to 50.

calcRedDim

Logical indicating whether to compute reduced dimensions (PCA, UMAP, TSNE, UMAP2D, TSNE2D). Defaults to TRUE. If its set to FALSE but there is no 2D nor 3D or not at all reduced dimension calculated in the input: it calculates all. Instead if there is no 2D, it calculates PCA, UMAP2D, TSNE2D. And lastly if there is no 3D, it calculates PCA, UMAP, TSNE.

paramFindMarkers

List of parameters to pass to findMarkers to compute marker genes for clusters. Defaults to list(test.type="wilcox", pval.type="all", direction="any").

calcAllPartitions

Logical indicating whether to force the computation of markers and DEGs from the entire list of partitionVars. Defaults to FALSE and only partitions from partitionVars with less or equal to 30 levels will be account for markers and DEGs calculations.

cells2keep

List of names for cells to keep when subsampling data. NOTE: Subsampling is only used in large datasets for visual purposes, and it does not affect computations. Only 50k cells will be used for visualization in the app and their indexs are stored in the CELLS2KEEP element of the output list.

descriptionText

Optional short description of the object being analized to be displayed in the scX app. This can help when working with multiple tabs.

verbose

Logical for step by step status while function is running. Defaults to TRUE.

Value

A named List that serves as input for launch_scX which contains the fields:

SCE:

SingleCellExperiment object with a computed normalized expression, dimensional reduction embedding (PCA, UMAP, TSNE, in 2D & 3D) calculated using the list of chosen.hvg if not NULL or the top nHVGs. colData contains the partitionVArs and metadataVars if they where specified, if not only a quick clusterization will be available for preliminar analyisis of the data.

sce.degs:

A named List of DataFrames where each DataFrame contains the consolidated marker statistics for each gene (row) for the cluster of the same name. Statistics are computed using findMarkers and user can choose test.type parameters to pass to that function. See combineMarkers for the details of how these dataframes are confectioner.

sce.markers:

A List of named Lists of DataFrames. Each one corresponds to the marker genes of every cluster in a partition (names of the nested lists). summary.stats: AUC if test.type=="wilcox" and -log.FC for test.type=="t" or test.type=="binom". log.FDR: -Log.FDR of the most appropriate inter-cluster comparison according to the selected p-value type. See findMarkers for the details of how these metrics are computed. boxcor: Correlation scores between the normalized gene expression profiles and a binary vector of cells, in which cells of the selected cluster have a value of 1.

(optional) text:

String if descriptionText not NULL.

CELLS2KEEP:

Numeric, the indices of the selected cells chosen for visualization in the scX application. (Alternatively) Character, if 'all' no subsampling will be performed for visualization purposes.

Details

This function handle the basic preprocessing steps for sc/sn-RNAseq experiments data taking advantage of the utility power of scran, scater and SingleCellExperiment packages. The steps are: Converting the input to SingleCellExperiment object, calculate some QC metrics, normalization of the expression matrix, find the most highly variable genes, compute embeddings for various dimensionality reduction techniques, compute if not specified a clustering of the cells for marker genes and DEGs analysis, subsampling the SingleCellExperiment object to reduce waiting times between plots in the scX app. Below you can find more details for these steps that are recommended in OSCA book.

Partitions

It is extremely recommended that a curated partition of the data be present in the SingleCellExperiment object or in the metadata. Marker genes and differential expression analysis will be compute for partitions that are passed through partitionVars and have less than 31 groups. If user wants run the calculations even if some partitions have more than 30 levels must set calcAllPartitions to TRUE. It may be worth coloring some partitions in plots but without having to compute marker genes and DEGs analyses for them. Those partitions had to be passed through metadataVars. If partitionVars and metadataVars are both NULL a quick clustering (see "Normalization") is used for the marker genes and DEGs analysises.

Normalization

If there is no normalized assay in the SingleCellExperiment object or Seurat object: Perform a Normalization by deconvolution proposed in the OSCA book. First, calculate clusters using the Walktrap community detection algorithm for graph-based clustering with default parameters from quickCluster. The resulting clusters are stored in colData as "scx.clust". Then compute scale factors for the cells using those clusters. Finally, calculate the lognormalized expression matrix by applying a log2 transformation to the product of the raw matrix and scale factors with pseudocounts of 1. If a normalized assay exists and "scx.clust" is included in the partitionVars, the function described before will be applied to compute the clusters, which are stored in colData.

Highly Variable Genes

If chosen.hvg is not specified, it will calculate the top nHVGs most variable genes with biological component > 0 This is computed with modelGeneVar which calculates the variance and mean of the lognormalized expression values. By fitting a trend of the variance against the mean, a biological component of variation for each gene can be assigned as the residual from the trend. See appropiate documentation for more details.

Reduced Dimensional Analysis Techniques

scX is a tool that helps visualizing the data properly, so it runs some dimensionality reduction analysis technique (PCA, UMAP2D, TSNE2D, UMAP3D, TSNE3D) If xx already has dimensionality reduction embeddings calculated, calcRedDim can be set to FALSE but:

if there is no dimRed calculated in the SingleCellExperiment object -> it calculates all

if there is no 2D or 3D redDim caclulated -> it calculates all

if there is no 2D -> it calculates only 2D

if there is no 3D -> it calculates only 3D Principal Component Analysis is calculated with runPCA using the chosen.hvg and retaining the first nPCs components, for the normalized expression matrix. TSNE and UMAP are calculated with runTSNE and runUMAP functions using the PCA matrix.

Marker Genes Analysis

For all partitions in partitionVars createSCEobject compute statistics for identify marker genes for every cluster using findMarkers with parameters specified by the user in paramFindMarkers. Only genes with FDR<0.05 are selected for each cluster, and the boxcor is calculated for those genes.

boxcor:

The boxcor is the correlation between a gene's expression vector (logcounts) and a binary vector, where only the cells from the selected cluster mark 1 while the rest of the cells mark 0.

Differential Expression Analysis

For all partitions in partitionVars createSCEobject compute statistics for identify DEGs between clusters using findMarkers by setting direction="any" and pval.type="all", the test.type is specified by the user in paramFindMarkers, with the exception that if the user had specified "wilcox" as test.type, the logFC values would be extracted from the function with test.type="t".

Subsampling cells

If the SingleCellExperiment object contains over 50k cells, a random sample of 50k cells will be chosen for visualization in the application. Cell names that the user wants to keep in the visualizations can be passed to cells2keep. Please note that all calculations are already completed, and this step is solely for promoting smooth and efficient visualization.

See also

Author

Tomás Vega Waichman, Maria Luz Vercesi, Ariel A. Berardino, Maximiliano S. Beckel, Chernomoretz Lab and Collaborators

Examples

if (FALSE) {
# Quick start guide example:
library(scX)
cseo <- createSCEobject(xx = sce, 
                        partitionVars = "inferred_cell_type", 
                        metadataVars = c("source_name", "age", "sex", "strain", "treatment", "pseudotime"),
                        descriptionText = "Quick Start Guide")
launch_scX(cseo)
}