Creates an object ready for use in scX

createSCEobject creates the input object (a List) for the function launch_scX. This list includes a SingleCellExperiment object with normalized expression, reduced dimensions for visualization, and any additional data provided. Also, this object includes identified gene markers and differential expression analysis for each user-specified partition (if no partition was selected, a quick clusterization is computed and considered)

Usage

createSCEobject(
  xx,
  assay.name.raw = "counts",
  assay.name.normalization = "logcounts",
  metadata = NULL,
  partitionVars = NULL,
  metadataVars = NULL,
  chosen.hvg = NULL,
  nHVGs = 3000,
  nPCs = 50,
  calcRedDim = TRUE,
  paramFindMarkers = list(test.type = "wilcox", pval.type = "all", direction = "up"),
  calcAllPartitions = FALSE,
  cells2keep = NULL,
  descriptionText = NULL,
  verbose = TRUE
)

Arguments

xx

Either a numeric count matrix object (with genes as rows and cells as columns), a SingleCellExperiment object or a Seurat object.

assay.name.raw

Assay name for the raw counts matrix, if object is a SingleCellExperiment. Defaults to counts.

assay.name.normalization

Assay name for normalized matrix if present in SingleCellExperiment. If not present, it computes logcounts (default).

metadata

(Optional) DataFrame with cell metadata. Rownames of the metadata data frame must include all cell names appearing as columns in xx.

partitionVars

(Optional) Names of metadata or colData columns used for gene markers and differential expression analysis. If NULL (default), a quick clusterization, using quickCluster from scran package will be computed.

metadataVars

(Optional) Names of additional metadata or colData columns to use for coloring in plots. If NULL (default), only partitionVars columns will be available for coloring plots.

chosen.hvg

(Optional) list of Highly Variable Genes. NOTE: if chosen.hvg=NULL and xx is a Seurat object with computed VariableFeatures then this parameter will be set to that list of genes.

nHVGs

Number of Highly Variable Genes to use if chosen.hvg=NULL. Defaults to 3000.

nPCs

Number of Principal Components to use in PCA. Defaults to 50.

calcRedDim

Logical. Indicates whether to compute reduced dimensions (PCA, UMAP, TSNE, UMAP2D, TSNE2D). Defaults to TRUE. (Note: If it is set to FALSE but there is no 2D reduced dimension provided in the input, 2-dim PCA, UMAP and TSNE embeddings will be estimated. The same happens for 3D embeddings).

paramFindMarkers

named List of parameters to pass to findMarkers to compute cluster gene markers. Defaults to list(test.type="wilcox", pval.type="all", direction="any").

calcAllPartitions

Logical. Defaults to FALSE, meaning that only partitions from partitionVars with less or equal to 30 levels will be considered for markers and DEGs calculations. If TRUE, force the computation of markers and DEGs for the entire list of partitionVars.

cells2keep

(Optional) List of cell names to keep in case of subsampling. NOTE: Subsampling is only activated for visual purposes for large datasets, it is not used for computations. 50k cells will be used for visualization in the app. Their indexes are stored in the CELLS2KEEP element of the CSEO object.

descriptionText

(Optional) Short description of the object being analized. This text will be displayed in the Summary module of the scX app.

verbose

Logical for step by step status while function is running. Defaults to TRUE.

Value

A named List that serves as input for launch_scX which contains the fields:

SCE:

SingleCellExperiment object with a computed normalized expression, dimensional reduction embedding (PCA, UMAP, TSNE, in 2D & 3D) calculated using the list of chosen.hvg if not NULL or the top nHVGs. colData contains the partitionVArs and metadataVars if they where specified, if not only a quick clusterization will be available for preliminar analysis of the data.

sce.degs:

A named List of DataFrames where each DataFrame contains the consolidated marker statistics for each gene (row) for the cluster of the same name. Statistics are computed using findMarkers and user can choose test.type parameters to pass to that function. See combineMarkers for the details of how these dataframes are confectioned.

sce.markers:

A List of named Lists of DataFrames. Each one corresponds to the marker genes of every cluster in a partition (names of the nested lists). summary.stats: AUC if test.type=="wilcox" and -log.FC for test.type=="t" or test.type=="binom". log.FDR: -Log.FDR of the most appropriate inter-cluster comparison according to the selected p-value type. See findMarkers for the details of how these metrics are computed. boxcor: Correlation scores between the normalized gene expression profiles and a binary vector of cells, in which cells of the selected cluster have a value of 1.

(optional) text:

String if descriptionText not NULL.

CELLS2KEEP:

Numeric, the indices of the selected cells chosen for visualization in the scX application. (Alternatively) Character, if 'all' no subsampling will be performed for visualization purposes.

Details

This function handles the basic preprocessing steps for sc/sn-RNAseq data. It leverages on functionality implemented in scran, scater and SingleCellExperiment packages. The steps are: Converting the input to SingleCellExperiment object, the estimation of QC metrics, the normalization of the expression matrix, the identification of highly variable genes, the computation of embeddings for various dimensionality reduction techniques. In case no cell partition is provided, a default clustering step is conducted in order to investigate marker genes and Differential expression patterns. For large datasets (number of cells > 50k) a subsampling of the SingleCellExperiment object is considered to reduce waiting times in the scX app. A thorough discussion of the employed preprocessing pipeline can be found at OSCA book.

Partitions

It is recommended to include a curated partition of the data in the input object or in the metadata. Marker genes and differential expression analysis automatically will be computed for partitions having less than 31 levels, that are specified as partitionVars. If user wants to run the calculations for partitions presenting more than 30 levels calcAllPartitions must be set to TRUE (Note that this step could be very time consuming). It may be worth coloring some partitions in plots but without having to compute marker genes and DEGs analyses for them. Those partitions must be passed through metadataVars. If partitionVars and metadataVars are both NULL a quickclustering step (see "Normalization") is used in order to identify markers and DEGs.

Normalization

If there is no normalized assay in the SingleCellExperiment object or Seurat input object a Normalization by deconvolution is performed as proposed in the OSCA book. First, we calculate clusters using the Walktrap community detection algorithm for graph-based clustering (default parameters from quickCluster). The resulting clusters are stored in colData as "scx.clust". Then we compute scale factors for the cells using those clusters. Finally, we calculate the lognormalized expression matrix by applying a log2 transformation to the product of the raw matrix and scale factors considering a pseudocounts of 1. If a normalized assay exists and "scx.clust" is included in the partitionVars, the function described before will be applied to compute the clusters, which are stored in colData.

Highly Variable Genes

If chosen.hvg is not specified, we will use modelGeneVar to calculate the variance and mean of the lognormalized expression values. By fitting a trend of the variance against the mean, a biological component of variation for each gene can be assigned as the residual from the trend (see scran documentation for more details). We consider the top nHVGs most variable genes with biological component > 0.

Reduced Dimensional Analysis Techniques

scX is a tool that helps visualizing the data properly, so it runs some dimensionality reduction analysis techniques (PCA, UMAP2D, TSNE2D, UMAP3D, TSNE3D). If xx already has dimensionality reduction embeddings calculated, calcRedDim can be set to FALSE but:

if there is no dimRed calculated in the SingleCellExperiment object -> it calculates all

if there is no 2D or 3D redDim caclulated -> it calculates all

if there is no 2D -> it calculates only 2D

if there is no 3D -> it calculates only 3D

Principal Component Analysis is calculated with runPCA using the chosen.hvg and retaining the first nPCs components, for the normalized expression matrix. TSNE and UMAP are calculated with runTSNE and runUMAP functions using the PCA matrix.

Marker Genes Analysis

createSCEobject computes statistics to identify marker genes for every cluster for all partitions in partitionVars using findMarkers functions (with parameters specified by the user in paramFindMarkers). Only genes with FDR<0.05 are selected for each cluster. The boxcor score is calculated for those genes as follows:

boxcor:

The boxcor is the correlation between a gene's expression vector (logcounts) and a binary vector, where only the cells from the selected cluster mark 1 while the rest of the cells mark 0.

Differential Expression Analysis

We use the findMarkers function to identify DEGs between clusters specified in partitionVars (direction="any",pval.type="all"). The test.type can be specified by the user in paramFindMarkers$test.type.

Subsampling cells

If the SingleCellExperiment object contains over 50k cells, a random sample of 50k cells will be chosen for visualization purposes in the application. Cell names that the user wants to keep in the visualizations can be specified in the cells2keep parameter. Please note that it is solely for producing efficient visualizations.

See also

Author

Tomás Vega Waichman, Maria Luz Vercesi, Ariel A. Berardino, Maximiliano S. Beckel, Chernomoretz Lab and Collaborators

Examples

if (FALSE) {
# Quick start guide example:
library(scX)
cseo <- createSCEobject(xx = sce, 
                        partitionVars = "inferred_cell_type", 
                        metadataVars = c("source_name", "age", "sex", "strain", "treatment", "pseudotime"),
                        descriptionText = "Quick Start Guide")
launch_scX(cseo)
}