SingleCellExperiment object ready for use in scXplorer

createSCEobject() returns a list which includes a SingleCellExperiment object with counts, at least one clusterization, gene markers for each clusterization, reduced dimensions for visualization, and any additional data provided. This list is used as input when launching the scXplorer app.

Usage

createSCEobject(
  xx,
  assay.name.raw = "counts",
  assay.name.normalization = "logcounts",
  metadata = NULL,
  toFactors = NULL,
  toKeep = NULL,
  chosen.hvg = NULL,
  nHVGs = 3000,
  nPCs = 50,
  calcRedDim = TRUE,
  paramFindMarkers = list(test.type = "wilcox", pval.type = "all", direction = "any"),
  calcAllToFactors = FALSE,
  cells2keep = NULL,
  descriptionText = NULL,
  verbose = TRUE
)

Arguments

xx

Either a matrix with counts, or a SCE or Seurat object.

assay.name.raw

Assay name for raw counts matrix if object is a SCE. Defaults to counts.

assay.name.normalization

Assay name for normalized matrix if present in SCE. If not present, it computes logcounts (default).

metadata

Optional dataframe with cell metadata.

toFactors

Optional metadata column names (or colData) to use as factors. If NULL (default), a quick clusterization will be computed.

toKeep

Additional metadata column names (or colData) to use only for coloring in plots. If NULL (default), only toFactors columns will be available for coloring plots.

chosen.hvg

Optional list of Highly Variable Genes.

nHVGs

Number of Highly Variable Genes to use if chosen.hvg=NULL. Defaults to 3000.

nPCs

Number of Principal Components to use in PCA. Defaults to 50.

calcRedDim

Logical indicating whether to compute reduced dimensions (PCA, UMAP, TSNE, UMAP2D, TSNE2D). Defaults to TRUE.

paramFindMarkers

List of parameters to pass to scran::findMarkers(...) to compute marker genes for clusters. Defaults to list(test.type="wilcox", pval.type="all", direction="any").

calcAllToFactors

Logical indicating whether to force the computation of markers and DEGs from the entire list of toFactors.

cells2keep

List of names for cells to keep when subsampling data. NOTE: Subsampling is only used in large datasets for visual purposes, and it does not affect computations.

descriptionText

Optional short description of the object being analized. This can help when working with multiple tabs.

verbose

Logical for step by step status while function is running. Defaults to TRUE.

Value

List with a SingleCellExperiment object and additional data ready for use in scXplorer.